+++ Package.swift
@@ -14,8 +14,9 @@ func envFlag(_ name: String) -> Bool {
     getenv(name).map { String(cString: $0) == "1" } ?? false
 }
 
-let enableKSCrash = envFlag("ENABLE_KSCRASH")
 let enableV10 = envFlag("SDK_V10")
+let v10SwiftSettings: [SwiftSetting] = enableV10 ? [.define("SDK_V10")] : []
+let v10CSettings: [CSetting] = enableV10 ? [.define("SDK_V10", to: "1")] : []
 
 var products: [Product] = [
     .library(name: "SentryDistribution", targets: ["SentryDistribution"])
@@ -105,17 +106,16 @@ if enableV10 {
 } else {
     products.append(.library(name: "SentrySPM", targets: ["SentryObjCInternal"]))
 }
-products.append(.library(name: "Sentry+KSCrash", targets: ["SentryObjCInternal"]))
-
 let sentrySwiftTarget: Target = .target(
     name: "SentrySwift",
     dependencies: ["_SentryPrivate", "SentryHeaders"],
     path: "Sources/Swift",
+    cSettings: v10CSettings,
+    swiftSettings: v10SwiftSettings
 )
 
-if enableKSCrash {
+if enableV10 {
     sentrySwiftTarget.dependencies.append(.product(name: "Installations", package: "KSCrash"))
-    sentrySwiftTarget.swiftSettings?.append(.define("ENABLE_KSCRASH"))
 }
 
 targets += [
@@ -124,14 +124,17 @@ targets += [
         name: "SentryHeaders",
         path: "Sources/Sentry",
         sources: ["SentryDummyPublicEmptyClass.m"],
-        publicHeadersPath: "Public"
+        publicHeadersPath: "Public",
+        cSettings: v10CSettings
     ),
     .target(
         name: "_SentryPrivate",
         dependencies: ["SentryHeaders"],
         path: "Sources/Sentry",
         sources: ["SentryDummyPrivateEmptyClass.m"],
-        publicHeadersPath: "include"),
+        publicHeadersPath: "include",
+        cSettings: v10CSettings
+    ),
 
     sentrySwiftTarget,
 
@@ -161,7 +164,8 @@ targets += [
             .headerSearchPath("SentryCrash/Recording/Tools"),
             .headerSearchPath("SentryCrash/Installations"),
             .headerSearchPath("SentryCrash/Reporting/Filters"),
-            .headerSearchPath("SentryCrash/Reporting/Filters/Tools")])
+            .headerSearchPath("SentryCrash/Reporting/Filters/Tools")
+        ] + v10CSettings)
 ]
 
 // BEGIN:OBJC_WRAPPER
@@ -170,7 +174,10 @@ targets += [
     .target(
         name: "SentryObjCCompat",
         dependencies: ["SentryObjCInternal"],
-        path: "Sources/SentryObjCCompat"),
+        path: "Sources/SentryObjCCompat",
+        cSettings: v10CSettings,
+        swiftSettings: v10SwiftSettings
+    ),
     .target(
         name: "SentryObjC",
         dependencies: ["SentryObjCCompat"],
@@ -178,12 +185,12 @@ targets += [
         publicHeadersPath: "Public",
         cSettings: [
             .headerSearchPath("Public")
-        ]
+        ] + v10CSettings
     )
 ]
 // END:OBJC_WRAPPER
 
-let packageDependencies: [Package.Dependency] = enableKSCrash ? [.package(url: "https://github.com/kstenerud/KSCrash.git", from: "2.6.0-beta.5")] : []
+let packageDependencies: [Package.Dependency] = enableV10 ? [.package(url: "https://github.com/kstenerud/KSCrash.git", from: "2.6.0-beta.5")] : []
 
 let package = Package(
     name: "Sentry",